requirements.txt
django~=5.1.4
djangorestframework~=3.15.2
django-cors-headers
requests~=2.32.3
polyline~=2.0.2
pyrebase4
setuptools
folium==0.12
geopy
Pillow